refactor: 유틸함수 리팩토링
