+++ src/config/index.ts
@@ -1,5 +1,2 @@
 // 환경 설정 export
-export { config } from './env';
-
-// 기본 export (가장 많이 사용되는 것)
-export { config as default } from './env';
+export * from './env';
