+++ src/constants/index.ts
@@ -1,6 +1,9 @@
 // API 관련
 export * from './api';
 
+// 인증 관련
+export * from './auth';
+
 // 게시판 관련
 export * from './board';
