+++ src/utils/data.ts
@@ -1,5 +1,3 @@
-import { getBreakpointValue } from '@/constants/ui';
-
 /**
  * 딜레이 함수 (Promise 기반)
  */
@@ -14,56 +12,12 @@ export const debounce = <T extends (...args: any[]) => any>(
   func: T,
   wait: number,
 ): ((...args: Parameters<T>) => void) => {
-  let timeout: NodeJS.Timeout;
+  let timeout: ReturnType<typeof setTimeout>;
 
   return (...args: Parameters<T>) => {
     clearTimeout(timeout);
     timeout = setTimeout(() => func(...args), wait);
   };
-};
-
-/**
- * 클립보드에 텍스트 복사
- */
-export const copyToClipboard = async (text: string): Promise<boolean> => {
-  // 입력값 검증
-  if (!text || typeof text !== 'string') {
-    console.warn('copyToClipboard: 유효하지 않은 텍스트');
-    return false;
-  }
-
-  try {
-    // 최신 Clipboard API 사용 (HTTPS 필요)
-    if (navigator.clipboard && window.isSecureContext) {
-      await navigator.clipboard.writeText(text);
-      return true;
-    } else {
-      // 폴백: document.execCommand 사용
-      const textArea = document.createElement('textarea');
-      textArea.value = text;
-      textArea.style.position = 'fixed';
-      textArea.style.left = '-999999px';
-      textArea.style.top = '-999999px';
-      textArea.style.opacity = '0';
-
-      document.body.appendChild(textArea);
-      textArea.focus();
-      textArea.select();
-
-      try {
-        const result = document.execCommand('copy');
-        document.body.removeChild(textArea);
-        return result;
-      } catch (execError) {
-        console.error('execCommand 복사 실패:', execError);
-        document.body.removeChild(textArea);
-        return false;
-      }
-    }
-  } catch (error) {
-    console.error('클립보드 복사 실패:', error);
-    return false;
-  }
 };
 
 /**
@@ -76,15 +30,16 @@ export const generateId = (length: number = 8): string => {
   }
 
   // 커스텀 길이를 위한 crypto.getRandomValues() 사용
-  const chars = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
+  const chars =
+    'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
   const array = new Uint8Array(length);
   crypto.getRandomValues(array);
-  
+
   let result = '';
   for (let i = 0; i < length; i++) {
     result += chars[array[i] % chars.length];
   }
-  
+
   return result;
 };
 
@@ -183,36 +138,4 @@ export const randomBetween = (min: number, max: number): number => {
     throw new Error('min은 max보다 작거나 같아야 합니다.');
   }
   return Math.floor(Math.random() * (max - min + 1)) + min;
-};
-
-/**
- * 화면 크기 확인
- */
-export const getScreenSize = () => {
-  const width = window.innerWidth;
-  const mdBreakpoint = getBreakpointValue('MD');
-  const lgBreakpoint = getBreakpointValue('LG');
-
-  return {
-    isMobile: width < mdBreakpoint,
-    isTablet: width >= mdBreakpoint && width < lgBreakpoint,
-    isDesktop: width >= lgBreakpoint,
-  };
-};
-
-/**
- * 온라인 상태 확인
- */
-export const isOnline = (): boolean => {
-  return navigator.onLine;
-};
-
-/**
- * 스크롤 맨 위로
- */
-export const scrollToTop = (): void => {
-  window.scrollTo({
-    top: 0,
-    behavior: 'smooth',
-  });
 };
